Toggle link instead of replacing selection
The previous code was replacing the selected text with a new link node,
but this removes any other inline markup decoration such as bold or
italic.

Instead of doing that, we could use the toggleLink helper from lixical,
which will keep the other markup in place, so we can combine bold,
italic and links.